Switch environment during process creation
When creating processes with 'CreateProcessWithLogonW', the secondary
logon service creates the correct environment for the corresponding user
automatically. For 'CreateProcessAsUserA' and 'CreateProcessWithTokenW'
this is not true. When using 'null' for 'lpEnvironment', new processes
are created with the environment block of the caller.

This matters, because certain applications require an properly setup
environment to work correcrly. I experienced this on some servers,
where PowerShell was not able to launch with a malformed environment.

With this commit RunasCs tries to obtain the EnvironmentBlock for the
desired user and uses this in the corresponding CreateProcess calls.
Additionally, it sets the 'USERPROFILE' variable manually, as this is
only set during a call to LoadProfile, which can only be called by
high privileged user accounts.