Switch to exception based error handling
The eror handling of RunasCs included calls to System.Environment.Exit.
This is fine when using the .exe, but when loading the assembly from
powershell a call to System.Environment.Exit causes PowerShell to quit.
In order to prevent this, an execption based error handling is
introduced with this commit. This way, also PowerShell should now stay
alive in all error cases.