Fix timeout for processes returning no output
The implementation of the timeout paramater was not fail safe. For
processes that do not return any output, the timeout was not working.
This was because the pipe is created in PIPE_WAIT mode and ReadFile
blocks if no data is present.

This is now fixed by calling SetNamedPipeHandleState and setting the
read pipe handle to PIPE_NOWAIT.